+++ uk.ac.gda.devices.xspress3/src/uk/ac/gda/devices/detector/xspress3/fullCalculations/Xspress3WithFullCalculationsDetector.java
@@ -110,7 +110,7 @@ public class Xspress3WithFullCalculationsDetector extends DetectorBase implement
 	}
 
 	@Override
-	public double[][] getMCData(double time) throws DeviceException {
+	public int[][] getMCData(double time) throws DeviceException {
 		return dataOperations.getMCData(time);
 	}
